Rename io submodule to fs (closes #26)
