+++ source/cppassist/include/cppassist/fs/DescriptiveRawFile.h
