+++ source/cppassist/include/cppassist/fs/RawFile.h
