+++ source/cppassist/include/cppassist/fs/SystemInfo.h
